Improved IDE (di) command.
OLD: if your segment definition.py module produced an .ly file that would
result in no PDF, then the IDE raised an exception.

NEW: if your segment definition.py module produces an .ly file that will result
in no PDF, then the IDE writes the .ly file to disk and conrol returns to the
composer and execution continues correctly.

+++ scoremanager/idetools/SegmentPackageManager.py
@@ -238,13 +238,15 @@ class SegmentPackageManager(ScoreInternalPackageManager):
                 self._io_manager._display_errors(stderr_lines)
                 return
             if not os.path.exists(illustration_pdf_path):
-                shutil.move(candidate_pdf_path, illustration_pdf_path)
-                shutil.move(candidate_ly_path, illustration_ly_path)
-                tab = self._io_manager._tab
                 messages = []
                 messages.append('Wrote ...')
-                messages.append(tab + illustration_ly_path)
-                messages.append(tab + illustration_pdf_path)
+                tab = self._io_manager._tab
+                if os.path.exists(candidate_ly_path):
+                    shutil.move(candidate_ly_path, illustration_ly_path)
+                    messages.append(tab + illustration_ly_path)
+                if os.path.exists(candidate_pdf_path):
+                    shutil.move(candidate_pdf_path, illustration_pdf_path)
+                    messages.append(tab + illustration_pdf_path)
                 self._io_manager._display(messages)
             else:
                 result = systemtools.TestManager.compare_files(
